# backend/swagger/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Test Tracker API",
    "version": "1.0.0",
    "description": "Documentazione completa delle API del progetto Test Tracker per supportare il frontend e il test delle route. Le API protette richiedono il token JWT nell'header Authorization come Bearer token."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Server locale di sviluppo"
    }
  ],
...
  "paths": {
...
    "/admin/users": {
      "get": {
        "tags": [
          "Main"
        ],
        "summary": "Redirect to admin users",
        "description": "Superadmin protected: redirects to the admin users management page.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "302": {
            "description": "Redirect to /admin/users",
            "headers": {
              "Location": {
                "description": "URL to redirect to",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
    "/api/users": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Elenco utenti",
        "description": "Restituisce la lista utenti. Per accedere è necessario inviare il token JWT nell'header Authorization come Bearer token.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista utenti disponibile",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/users/me": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Profilo utente autenticato",
        "description": "Restituisce i dati dell'utente autenticato tramite token JWT inviato nell'header Authorization come Bearer token.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Profilo utente restituito correttamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "description": "Token mancante o non valido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                },
                "examples": {
                  "missingToken": {
                    "value": {
                      "message": "Token mancante o formato non valido"
                    }
                  },
                  "invalidToken": {
                    "value": {
                      "message": "Token non valido"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Utente non trovato",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                },
                "example": {
                  "message": "Utente non trovato"
                }
              }
            }
          },
          "500": {
            "description": "Errore del server",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorWithDetail"
                },
                "example": {
                  "message": "Errore del server",
                  "specific": "Internal error"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "Aggiorna profilo utente autenticato",
        "description": "Aggiorna i dati dell'utente autenticato. Richiede il token JWT nell'header Authorization come Bearer token.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nome": {
                    "type": "string",
                    "example": "Mario"
                  },
                  "cognome": {
                    "type": "string",
                    "example": "Rossi"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "mario.rossi@example.com"
                  }
                },
                "required": [
                  "nome",
                  "cognome",
                  "email"
                ]
              },
              "example": {
                "nome": "Mario",
                "cognome": "Rossi",
                "email": "mario.rossi@example.com"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profilo aggiornato con successo",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessMessage"
                },
                "example": {
                  "message": "Profilo aggiornato con successo"
                }
              }
            }
          },
          "400": {
            "description": "Dati di input non validi o incompleti",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                },
                "example": {
                  "message": "Tutti i campi sono obbligatori"
                }
              }
            }
          },
          "401": {
            "description": "Token mancante o non valido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                },
                "examples": {
                  "missingToken": {
                    "value": {
                      "message": "Token mancante o formato non valido"
                    }
                  },
                  "invalidToken": {
                    "value": {
                      "message": "Token non valido"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Utente non trovato",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                },
                "example": {
                  "message": "Utente non trovato"
                }
              }
            }
          },
          "500": {
            "description": "Errore del server",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorWithDetail"
                },
                "example": {
                  "message": "Errore del server",
                  "specific": "Internal error"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/me/password": {
      "patch": {
        "tags": [
          "Users"
        ],
        "summary": "Aggiorna password utente autenticato",
        "description": "Aggiorna la password dell'utente autenticato. Richiede il token JWT nell'header Authorization come Bearer token.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "Oldpassword": {
                    "type": "string",
                    "format": "password",
                    "example": "PasswordVecchia123!"
                  },
                  "Newpassword": {
                    "type": "string",
                    "format": "password",
                    "example": "PasswordNuova123!"
                  }
                },
                "required": [
                  "Oldpassword",
                  "Newpassword"
                ]
              },
              "example": {
                "Oldpassword": "PasswordVecchia123!",
                "Newpassword": "PasswordNuova123!"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password aggiornata con successo",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessMessage"
                },
                "example": {
                  "message": "Password aggiornata con successo"
                }
              }
            }
          },
          "400": {
            "description": "Password attuale errata",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                },
                "example": {
                  "message": "Password errata"
                }
              }
            }
          },
          "401": {
            "description": "Token mancante o non valido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                },
                "examples": {
                  "missingToken": {
                    "value": {
                      "message": "Token mancante o formato non valido"
                    }
                  },
                  "invalidToken": {
                    "value": {
                      "message": "Token non valido"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Utente non trovato",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                },
                "example": {
                  "message": "Utente non trovato"
                }
              }
            }
          },
          "500": {
            "description": "Errore del server",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorWithDetail"
                },
                "example": {
                  "message": "Errore del server",
                  "specific": "Internal error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/users/{id}/relations": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Progetti e sessioni collegati a un utente",
        "description": "Restituisce i dati dell'utente insieme all'elenco dei progetti a cui è collegato (come tester assegnato o come responsabile) e alle sue sessioni di test, usato dalla pagina di dettaglio utente lato superadmin. Richiede il token JWT nell'header Authorization come Bearer token. Solo superadmin.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID dell'utente",
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Progetti e sessioni restituiti correttamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "$ref": "#/components/schemas/User"
                    },
                    "projects": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "integer" },
                          "name": { "type": "string" },
                          "status": { "type": "string" },
                          "relation": {
                            "type": "string",
                            "enum": ["Responsabile", "Tester"],
                            "description": "Ruolo dell'utente su questo progetto: Responsabile se manager_id coincide, Tester se assegnato tramite project_assignment"
                          }
                        }
                      }
                    },
                    "sessions": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "integer" },
                          "project_id": { "type": "integer" },
                          "status": { "type": "string" },
                          "started_at": { "type": "string", "format": "date-time" },
                          "completed_at": { "type": "string", "format": "date-time", "nullable": true },
                          "project_name": { "type": "string" }
                        }
                      }
                    }
                  }
                },
                "example": {
                  "user": {
                    "id": 1,
                    "nome": "Mario",
                    "cognome": "Rossi",
                    "email": "mario.rossi@example.com",
                    "role": "user"
                  },
                  "projects": [
                    {
                      "id": 3,
                      "name": "Sito e-commerce",
                      "status": "Attivo",
                      "relation": "Tester"
                    }
                  ],
                  "sessions": [
                    {
                      "id": 12,
                      "project_id": 3,
                      "status": "Completata",
                      "started_at": "2026-06-20T09:00:00Z",
                      "completed_at": "2026-06-20T10:30:00Z",
                      "project_name": "Sito e-commerce"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Token mancante o non valido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                },
                "example": {
                  "message": "Token mancante o formato non valido"
                }
              }
            }
          },
          "403": {
            "description": "Accesso negato",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                },
                "example": {
                  "error": "Accesso negato"
                }
              }
            }
          },
          "404": {
            "description": "Utente non trovato",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                },
                "example": {
                  "message": "Utente non trovato"
                }
              }
            }
          },
          "500": {
            "description": "Errore del server",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorWithDetail"
                },
                "example": {
                  "message": "Errore del server",
                  "specific": "Internal error"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/{id}/role": {
      "patch": {
        "tags": [
          "Users"
        ],
        "summary": "Aggiorna ruolo utente per ID",
        "description": "Aggiorna il ruolo di un utente specifico tramite ID. Richiede il token JWT nell'header Authorization come Bearer token.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID dell'utente",
            "example": 1
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "role": {
                    "type": "string",
                    "enum": [
                      "user",
                      "admin",
                      "superadmin"
                    ],
                    "example": "admin"
                  }
                },
                "required": [
                  "role"
                ]
              },
              "example": {
                "role": "admin"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ruolo aggiornato con successo",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessMessage"
                },
                "example": {
                  "message": "Ruolo aggiornato con successo"
                }
              }
            }
          },
          "400": {
            "description": "Ruolo non valido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                },
                "example": {
                  "message": "Ruolo non valido"
                }
              }
            }
          },
          "401": {
            "description": "Token mancante o non valido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                },
                "examples": {
                  "missingToken": {
                    "value": {
                      "message": "Token mancante o formato non valido"
                    }
                  },
                  "invalidToken": {
                    "value": {
                      "message": "Token non valido"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Utente non trovato",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                },
                "example": {
                  "message": "Utente non trovato"
                }
              }
            }
          },
          "500": {
            "description": "Errore del server",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorWithDetail"
                },
                "example": {
                  "message": "Errore del server",
                  "specific": "Internal error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Inserire il token JWT come: Bearer <token>"
      }
    },
    "schemas": {
...
      "SuccessMessage": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "required": [
          "message"
        ]
      },
      "ErrorMessage": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "required": [
          "message"
        ]
      },
      "ErrorWithDetail": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "specific": {
            "type": "string"
          }
        },
        "required": [
          "message"
        ]
      },
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "nome": {
            "type": "string",
            "example": "Mario"
          },
          "cognome": {
            "type": "string",
            "example": "Rossi"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "mario.rossi@example.com"
          },
          "role": {
            "type": "string",
            "example": "user"
          }
        },
        "required": [
          "id",
          "nome",
          "cognome",
          "email",
          "role"
        ]
      },
...
    }
  }
}